requirements.txt
app-version==1.0.1
asgiref==3.2.7
async-timeout==3.0.1
azure-common==1.1.25
azure-storage-blob==2.1.0
azure-storage-common==2.1.0
certifi==2020.4.5.2
cffi==1.14.0
chardet==3.0.4
cryptography==2.9.2
Django==3.0.6
django-storages==1.9.1
djangorestframework==3.11.0
idna==2.9
notify==0.3.1
Pillow==7.1.2
pycparser==2.20
python-dateutil==2.8.1
pytz==2020.1
requests==2.23.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.9
validate-email==1.3
